Add Support for Querying Views.
Add support for Querying Views using the CDTViewQueryOperation. It fully implements
the view querying API. It is based off the CDTQueryFindDocumentsOperation, as a result it uses two blocks
to process the response data, one block to process the rows element, and one block to signify completion of
operation to the application.

Fixes #32